feat: map-surface filter button, theme control to bottom-left, clean attribution
Filters move out of the side panel/drawer into a floating button + popover
on the map (bottom-right, above locate/zoom), with an applied-count badge.
Country + light/dark controls move to the bottom-left corner beside the
legend. Attribution becomes a plain single-line credit flush in the
bottom-left corner (non-compact, no pill box) on both breakpoints.
