+++ src/components/map/MapLibreContainer.tsx
@@ -33,12 +33,12 @@ function GeolocateControl({ position }: { position: string }) {
   );
   return null;
 }
-// Attribution lives bottom-left on mobile so the bottom-right corner holds only
-// the geolocate/layers stack (their offsets assume nothing sits below them);
-// desktop keeps it bottom-right under the zoom controls.
+// Attribution renders as a plain single-line credit flush in the bottom-left
+// corner (styled in index.css) — non-compact so it never collapses into the
+// (i) toggle or floats mid-stack among the map buttons.
 function AttributionCtl({ position }: { position: 'bottom-left' | 'bottom-right' }) {
   useControl(
-    ({ mapLib }) => new (mapLib as typeof maplibregl).AttributionControl({}),
+    ({ mapLib }) => new (mapLib as typeof maplibregl).AttributionControl({ compact: false }),
     { position },
   );
   return null;
@@ -163,17 +163,7 @@ export const MapLibreView = forwardRef<MapLibreViewHandle, MapLibreViewProps>(
   function MapLibreView({ onMarkersReady, mapKey }, ref) {
   const mapRef = useRef<MapRef>(null);
 
-  // Match the app's lg breakpoint so the attribution corner follows the layout
-  const [isMobileLayout, setIsMobileLayout] = useState(
-    () => typeof window !== 'undefined' && window.matchMedia('(max-width: 1023px)').matches
-  );
-  useEffect(() => {
-    const mq = window.matchMedia('(max-width: 1023px)');
-    const onChange = (e: MediaQueryListEvent) => setIsMobileLayout(e.matches);
-    mq.addEventListener('change', onChange);
-    return () => mq.removeEventListener('change', onChange);
-  }, []);
-  const attribPosition = isMobileLayout ? 'bottom-left' as const : 'bottom-right' as const;
+  const attribPosition = 'bottom-left' as const;
   const [popupInfo, setPopupInfo] = useState<PopupInfo | null>(null);
   const [cursor, setCursor] = useState<string>('');
   const lastFlyToRef = useRef<number>(0);
