+++ src/index.css
@@ -277,36 +277,27 @@ input[type="range"]::-moz-range-thumb {
   font-family: 'DM Sans', system-ui, sans-serif;
 }
 
+/* Attribution — plain single-line credit flush in the bottom-left corner,
+   no pill box; readable over both light and dark basemaps via text-shadow. */
 .maplibregl-ctrl-attrib {
-  background: rgba(10, 10, 15, 0.95) !important;
-  color: #9ca3af !important;
-  font-size: 11px !important;
-  padding: 3px 8px !important;
-  border-radius: 6px !important;
+  background: transparent !important;
+  color: rgba(156, 163, 175, 0.85) !important;
+  font-size: 10px !important;
+  line-height: 1 !important;
+  padding: 0 !important;
   max-width: none !important;
-}
-
-/* Ensure attribution stays in bottom-right corner */
-.maplibregl-ctrl-attrib.maplibregl-compact {
-  min-height: auto !important;
-  padding: 2px 24px 2px 6px !important;
-}
-
-/* Mobile attribution - keep compact and in corner */
-@media (max-width: 640px) {
-  .maplibregl-ctrl-attrib {
-    font-size: 9px !important;
-    padding: 2px 6px !important;
-    max-width: 180px !important;
-  }
-
-  .maplibregl-ctrl-attrib.maplibregl-compact {
-    padding: 2px 20px 2px 4px !important;
-  }
+  white-space: nowrap !important;
+  text-shadow: 0 1px 2px rgba(0, 0, 0, 0.7);
 }
 
 .maplibregl-ctrl-attrib a {
+  color: inherit !important;
+  text-decoration: none !important;
+}
+
+.maplibregl-ctrl-attrib a:hover {
   color: #0080BC !important;
+  text-decoration: underline !important;
 }
 
 .maplibregl-ctrl-group {
@@ -395,9 +386,23 @@ input[type="range"]::-moz-range-thumb {
   right: 16px !important;
 }
 
-/* Map style control — inline above zoom + geolocate controls */
+/* Attribution container — flush bottom-left corner, under the theme control */
+.maplibregl-ctrl-bottom-left {
+  bottom: 6px !important;
+  left: 16px !important;
+}
+
+/* Map style control (country + light/dark) — bottom-left corner, above the
+   attribution line, beside the legend chip */
 .map-page .map-style-control {
-  bottom: 208px; /* 24px (zoom offset) + 172px (zoom group + geolocate group + gaps) + 12px gap */
+  bottom: 24px;
+  left: 16px;
+}
+
+/* Filter button — bottom-right, above geolocate + zoom.
+   24px (ctrl offset) + 76px (zoom group) + 10px margin + 40px (geolocate) + 12px gap */
+.map-page .map-filter-control {
+  bottom: 162px;
   right: 26px; /* 16px (ctrl container offset) + 10px (maplibregl-ctrl margin) */
 }
 
@@ -435,11 +440,11 @@ input[type="range"]::-moz-range-thumb {
     display: none;
   }
 
-  /* Country button + theme pill float above geolocate with a small gap.
-     Offset tracks --drawer-height: 12px gap + 48px geolocate group + 8px gap. */
+  /* Country button + theme pill sit in the bottom-LEFT corner, above the
+     attribution line. Offset tracks --drawer-height. */
   .map-page .map-style-control {
-    bottom: calc(var(--drawer-height, 80px) + 68px);
-    right: 22px; /* 12px (ctrl container offset) + 10px (maplibregl-ctrl margin) */
+    bottom: calc(var(--drawer-height, 80px) + 30px);
+    left: 12px;
   }
 
   .map-page .map-style-control .country-trigger {
@@ -451,10 +456,16 @@ input[type="range"]::-moz-range-thumb {
     width: 38px;
   }
 
-  /* Ensure map controls don't overlap with bottom sheet.
-     Offset tracks --drawer-height + 12px gap. */
+  /* Filter button floats above geolocate with a small gap.
+     Offset tracks --drawer-height: 12px gap + 40px geolocate + 8px gap. */
+  .map-page .map-filter-control {
+    bottom: calc(var(--drawer-height, 80px) + 72px);
+    right: 22px; /* 12px (ctrl container offset) + 10px (maplibregl-ctrl margin) */
+  }
+
+  /* Attribution line hugs the corner just above the bottom sheet. */
   .map-page .maplibregl-ctrl-bottom-left {
-    bottom: calc(var(--drawer-height, 80px) + 12px) !important;
+    bottom: calc(var(--drawer-height, 80px) + 8px) !important;
     left: 12px !important;
   }
